Require plugin grammars to be complete.
Before, I was using a Node trait to wrap every node in the AST. This feels like
too much at this point and I've decided to require that every grammar plugin has
to be complete, such that the abstraction of parsing that grammar is above the
AST Node level.

So, now we can use enums in the parser, and whatever we return will be where the
dynamic dispatch happens.