Evaluate across multiple parsers.
This also drops the full fledged Visitor pattern for a more concise
approach that I think will work fine for the time being. It may even
be that the complicated visiting happens in the library module
rather than each of the 'plugins'.

A few weak points I had to work around, especially around errors.