Something that at least works.
This seems convoluted and I'm sure there are places that can be
improved/excised. I like the use of the enum in the model, though and it's going
to work elegently with the serialize/deserialization of things.

This, combined with a step after deserialization of scopes to ensure that those
enums are populated. This requires the implementation of a trait in the scope,
which I believe can be derived, eventually.