More span tweaks, a modifying action, finally.
Next step is to find a way to restore Scope after mutation to the
original entity and mark that as dirty.

I understand how similar a lot of this is to an ECS, but I feel like my
needs are a little different. Though, it may be possible to do what I
want on top of an existing ECS.

At any rate, I won't know until I've used Rust enough and this is the
most I've written.