Thread local Session/Infrastructure.
I'm really unhappy with how things are working right now to introduce
this layer to the domain model. In a perfect world it wouldn't even be
necessary, but my hope is to hide the detail from users as much
as possible.

To that end, the code was way messier than necessary.

At the end of the day I was avoiding this approach because I felt like
it was messy, but it's better than what's happening now and I know
there are solutions that will work under tokio, so.... hey why not.

This will let me focus more on building the architecture I want.