Domain events published server-wide. Whew.
It's been a while and I've been working on this basically the whole time on and
off. I found this to be one of the most frustrating refactors so far. Going to
be distilling the lesson for a while, not totally sure why that's the case.

At any rate, events published from the domain layer end up queued in the
Session. When we close the Session they are observed by their audiences and then
flushed out to a Notifier. For web sockets this gets published to clients based
on their key.

Some things I'm unhappy with.

I'm kind of annoyed at the explosion of model structs in the reply layer. I'm
probably going to end up moving these dynamic json SimpleObservation's to
something more structured.

The web socket code is confusing. The only way I could find to support the
behavior I needed, which was being able to transmit replies as well as server
wide broadcasts. I'm sure this could both be cleaned up and improved.