It's time to rethink my approach.
What's starting to happen is the interplay between the scopes and combining
their mutability with that of the parent entity is causing things to get
spaghettified.

This is especially true of some of the areas of code that deal with lazy
loading. For example, the code in this commit needs to pass over all the
contained items to see if any of them have the same kind as the item being held.
Except, sometimes one of those items is the container that the item is being
removed from which will be immutably borrowed during the "take out"
action. I'm sure I could work around this, but I feel like the friction is
a sign to maybe rethink the approach.

I'm starting to appreciate the value of how some of the ECS' have solved this
problem, basically moving the scope and scope_mut calls to the Session or
something similar and handling things there.