New design fixes broken test!
I'm honestly kind of proud of how quickly I got here. What before was
kind of messy has gotten easier and shown to at least work for What
was broken in the old system.

I believe this means the next thing we do is remove EntityPtr's,
which is gonna suck haha