Fefactor towards Entry/Handle approach.
Wow, that was a wild ride!

Very first pass on migrating to the new Entry/Handle based approach that I
mocked up in the earlier hacking commits.

Things are still very much in flux but all the tests pass that were passing
before, which I'll take as a win. There's a _ton_ of needless cloning and
converting and looking up.

Just keep going, I guess.