Merge Session and DomainInfrastructure.
I feel like these two evolved back when things were organized differently enough
that this is worth revisiting. Now that the line is a little more clear, and
we've got Session implementing the Infrastructure trait it's easier and more
intuitive to pass Session. This will give us a good place to begin breaking up
the Session implementation.