I like this order better.
It would be so cool if I could generalize on SessionRef somehow. To do
that the calls to set/get would need to be generic on the type they
needed, and then I think the other parts wouldn't be so bad. It basically
amounts to how axum and other frameworks pass app data, so I should look
at how that's done.