Spawn blocking threads for Sessions.
Right now we're largely synchronous and most of the async code we have
is in CLI for setup and other work.

My problem is that for the RPC work I'd like to be able to use tokio
for that code, which means nesting runtimes so we need to break
free from the async thread we're running on. It's messy but when
async traits is stable we can probably pretty easily get rid of it by
moving async down a few layers.... maybe.