No more have_surroundings on Plugin.
Instead, the middleware can operate on/around Perform::Surroundings. This
is better for a number of reasons, the main one being before,
have_surroundings' purpose was pretty ill-defined. What are plugins
supposed to be doing in there and how does it relate t o what's happening
after. Now the relationship and options are more clear.