Don't send user their own Arrivals.
Ran into a bunch of odd stuff on this one. At first I wanted to have
an Audience variant that wrapped another and excluded a set of keys,
only that would crash pretty horribly in the Rune code and I'm too
tired to chase something like that right now.

I do think the idea is worth visiting, especially as I've been
thinking of changing how Audience works. For example always resolving
to keys at the time of raising.