Remove wasm example.
I'm doubting that wasm will find a place in this in the short term and the
example, plus having the wasm32 target dependency makes builds slow. There's
also a very good chance that we'll archive the wasm plugin and the rpc plugin in
the future for the same reason. This is a baby step towards that. Tagged main
before as archived/wasm/example.