Working state to/from behaviors scope.
The implementation of this is meh. It's annoying that I need the stub
to get a rune Struct for the state. I'm at least happy that filling in
the state is working. Just think lots could be cleaned up.

I tried just returning a Type but getting RTTI for a type hash is
private in rune.